Plot single feature seyr data

+++ XferHelper/Functions.fs
@@ -646,7 +646,7 @@ module Report =
         |> Array.filter (fun x -> x.Name = name)
         |> Array.map (fun x -> x.Score)
 
-    let rescoreFeature (data: Entry [], name: string, min: int, max: int) =
+    let rescoreFeature (data: Entry [], name: string, min: float, max: float) =
         for d in data do
             if d.Name = name then
                 if d.Score >= min && d.Score <= max then
